fix #1 & Move the scripts into l2u/
