+++ l2u/face.py
@@ -3,8 +3,8 @@
 # By Llf0703
 # 创建题面 problem.md 文件
 
-from file import get_dir
-from info import get_face
+from .file import get_dir
+from .info import get_face
 
 
 def face(pid):
